Solution: December 01 Language: JavaScript Tests: Passing
My goal is to tackle each problem in a unique programming language.  First up, JavaScript.

I'm recently doing more FP and wanted to solve this problem for all "n-ish" numbers.

My first approach was brute force and involved a HoF named "isNumberish" that called two other HoFs "isPowerOf" and "isSumOfUniquePowersOf".

isPowerOf went through a few iterations but landed on dividing the ceiling of the power of n by the floor of the power of n and seeing if they were equal to eachother.

I got stuck with isSumOfUniquePowersOf(y)(x).  I first tried summing 7^[0...x] and checking if at any point the sum equalled x.  This didn't return true for 8 or 50, so I realized I wasn't going about "unique powers" correctly.
I later had an aha moment with the "unique powers" and adapted function of O(n^2) to similarly find all subsums and see if any equalled n.

This strategy returned a lot of false positives, so I decided to reframe the issue and move on to finding sevenish numbers instead of validating them.

To find the nth sevenish number, I create a return array and iterate until that array has n numbers in it.  At first I looped over all natural numbers before realizing I could just loop over the powers of 7.

It was a bit of a cascade moment when I merged the previous approach so that in each iteration, I looked for subsums and merged them into the return array.  However, if at any point the length of the return array and the number of sums found within an iteration exceeded n, just return the latest sum.

If close to the end, the function returns the last element of the return array, which is always the last found power of 7.

Finally, by using currying to turn 7 into a variable, this approach solves for all n-ish sets.